functions/requirements.txt
face-recognition==1.3.0
Flask==2.3.3
flask-cors==4.0.0
Pillow==12.1.0
numpy==1.24.3
firebase-admin==6.2.0
requests==2.31.0
python-dotenv==1.0.0
